requirements.txt
absl-py
aio-pika
aiormq
astunparse
blis
cachetools
catalogue
catboost
certifi
cffi
chardet
click
cryptography
cycler
cymem
Cython
DAWG-Python
deeppavlov
deepspeed
docopt
fastai
fastapi
fastcore
fastprogress
fasttext
filelock
flashtext
flatbuffers
gast
gensim
google-auth
google-auth-oauthlib
google-pasta
graphviz
grpcio
h11
h5py
httptools
huggingface-hub
idna
intervaltree
ipymarkup
Jinja2
joblib
Keras
keras-nightly
Keras-Preprocessing
kiwisolver
lightgbm
Markdown
MarkupSafe
matplotlib
mpi4py
multidict
murmurhash
natasha
navec
ninja
nltk
numpy
oauthlib
onnx
onnxruntime
opt-einsum
overrides
packaging
pamqp
pandas
pathy
Pillow
plotly
preshed
prometheus-client
protobuf
psutil
pyasn1
pyasn1-modules
pybind11
pycparser
pydantic
pymorphy2
pymorphy2-dicts
pymorphy2-dicts-ru
pyOpenSSL
pyparsing
pyTelegramBotAPI
python-dateutil
pytz
PyYAML
razdel
regex
requests
requests-oauthlib
retrying
rsa
ruamel.yaml
rusenttokenize
sacremoses
scikit-learn
scipy
six
slovnet
smart-open
sortedcontainers
spacy
spacy-legacy
srsly
starlette
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorboardX
tensorflow-estimator
tensorflow-gpu
termcolor
thinc
threadpoolctl
tokenizers
torch
torchvision
tqdm
transformers
typer
typing-extensions
urllib3
uvicorn
wasabi
websockets
Werkzeug
wrapt
xgboost
yargy
yarl
